refactor: unify vendor package handling with generic function
